+++ client/src/commands.ts
@@ -1,44 +1,101 @@
 import * as path from 'path'
-import { commands, ExtensionContext, ShellExecution, Task, tasks, window, workspace } from 'vscode'
-
+import * as vscode from 'vscode'
+import {
+  commands,
+  ExtensionContext,
+  ShellExecution,
+  Task,
+  tasks,
+  window,
+  workspace,
+} from 'vscode'
+import {
+  asShellString,
+  fsToShell,
+  unknownToShell,
+} from './platform-string-utils'
 
 export const subscribeWollokCommands = (context: ExtensionContext): void => {
   context.subscriptions.push(registerCLICommand('wollok.start.repl', startRepl))
-  context.subscriptions.push(registerCLICommand('wollok.run.allTests', runAllTests))
+  context.subscriptions.push(
+    registerCLICommand('wollok.run.allTests', runAllTests),
+  )
   context.subscriptions.push(registerCLICommand('wollok.run.tests', runTests))
+  context.subscriptions.push(
+    registerCLICommand('wollok.run.program', runProgram),
+  )
 }
 
 /**
  * CLI Commands
  */
 
-const runTests = (filter: string) => wollokCLITask('run tests', 'Wollok run tests', ['test', `'${filter}'`])
+export const runProgram = (fqn: string): Task =>
+  wollokCLITask('run program', 'Wollok run program', [
+    'run',
+    `'${fqn}'`,
+    '--skipValidations',
+  ])
 
-const runAllTests = () => wollokCLITask('run tests', 'Wollok run all tests', ['test'])
+export const runTests = (filter: string): Task =>
+  wollokCLITask('run tests', 'Wollok run tests', [
+    'test',
+    `'${asShellString(filter)}'`,
+    '--skipValidations',
+  ])
 
-const startRepl = () => {
+export const runAllTests = (): Task =>
+  wollokCLITask('run tests', 'Wollok run all tests', [
+    'test',
+    '--skipValidations',
+  ])
+
+export const startRepl = (): Task => {
   const currentDocument = window.activeTextEditor.document
   const currentFileName = path.basename(currentDocument.uri.path)
-  return wollokCLITask('repl', `Wollok Repl: ${currentFileName}`, ['repl', currentDocument.fileName])
+
+  const replTask = wollokCLITask('repl', `Wollok Repl: ${currentFileName}`, [
+    'repl',
+    fsToShell(currentDocument.uri.fsPath),
+    '--skipValidations',
+  ])
+
+  setTimeout(() => {
+    vscode.commands.executeCommand('simpleBrowser.show', 'http://localhost:3000/')
+  }, 1000)
+
+  return replTask
 }
 
 /**
  * Helpers
  */
 
-const registerCLICommand = (command: string, taskBuilder: (...args: any[]) => Task) =>
-  commands.registerCommand(command, (...args) => tasks.executeTask(taskBuilder(args)))
+const registerCLICommand = (
+  command: string,
+  taskBuilder: (...args: any[]) => Task,
+) =>
+  commands.registerCommand(command, (...args) =>
+    tasks.executeTask(taskBuilder(args)),
+  )
 
 const wollokCLITask = (task: string, name: string, cliCommands: string[]) => {
-  const wollokCli = workspace.getConfiguration('wollokLinter').get('cli-path')
+  const wollokCli = unknownToShell(
+    workspace.getConfiguration('wollokLinter').get('cli-path'),
+  )
   const folder = workspace.workspaceFolders[0]
-  const shellCommand = [wollokCli, ...cliCommands].join(' ')
+  const shellCommand = [
+    wollokCli,
+    ...cliCommands,
+    '-p',
+    fsToShell(folder.uri.fsPath),
+  ].join(' ')
 
   return new Task(
     { type: 'wollok', task },
     folder,
     name,
     'wollok',
-    new ShellExecution(shellCommand)
+    new ShellExecution(shellCommand),
   )
 }