+++ client/src/extension.ts
@@ -4,20 +4,29 @@
  * ------------------------------------------------------------------------------------------ */
 
 import * as path from 'path'
-import { ExtensionContext, workspace, languages } from 'vscode'
-import { LanguageClient,
+import {
+  ExtensionContext,
+  workspace,
+  languages,
+  window,
+  StatusBarAlignment,
+} from 'vscode'
+import {
+  LanguageClient,
   LanguageClientOptions,
   ServerOptions,
-  TransportKind } from 'vscode-languageclient/node'
+  TransportKind,
+  WorkDoneProgress,
+} from 'vscode-languageclient/node'
 import { subscribeWollokCommands } from './commands'
-import { allWollokFiles } from './test/helper'
+import { allWollokFiles } from './utils'
 
 let client: LanguageClient
 
 export function activate(context: ExtensionContext): void {
   // The server is implemented in node
   const serverModule = context.asAbsolutePath(
-    path.join('server', 'out', 'server.js')
+    path.join('server', 'out', 'server.js'),
   )
   // The debug options for the server
   // --inspect=6009: runs the server in Node's Inspector mode so VS Code can attach to the server for debugging
@@ -53,16 +62,26 @@ export function activate(context: ExtensionContext): void {
     'wollok-lsp-ide',
     'Wollok',
     serverOptions,
-    clientOptions
+    clientOptions,
   )
 
   // Force first validation
   validateWorkspace()
 
-  // Force environment to restart
-  const revalidateWorskpace = _event =>
-    client.sendRequest('STRONG_FILES_CHANGED').then(validateWorkspace)
+  const statusBarItem = window.createStatusBarItem(StatusBarAlignment.Left)
 
+  client.onProgress(WorkDoneProgress.type, 'wollok-build', (progress) => {
+    if (progress.kind === 'begin' || progress.kind === 'report') {
+      statusBarItem.text = '$(loading~spin) Wollok Building...'
+      statusBarItem.show()
+    } else {
+      statusBarItem.hide()
+    }
+  })
+
+  // Force environment to restart
+  const revalidateWorskpace = (_event) =>
+    client.sendRequest('STRONG_FILES_CHANGED').then(validateWorkspace)
 
   workspace.onDidDeleteFiles(revalidateWorskpace)
   workspace.onDidRenameFiles(revalidateWorskpace)