+++ client/src/test/runTest.ts
@@ -16,7 +16,13 @@ async function main() {
     const extensionTestsPath = path.resolve(__dirname, './index')
 
     // Download VS Code, unzip it and run the integration test
-    await runTests({ extensionDevelopmentPath, extensionTestsPath })
+    await runTests({
+      extensionDevelopmentPath,
+      extensionTestsPath,
+      launchArgs: [
+        path.resolve(extensionDevelopmentPath, './client/testFixture'),
+      ],
+    })
   } catch (err) {
     console.error('Failed to run tests')
     process.exit(1)